test(cli): split logs suite
